CLEAN all sources
