+++ src/ModelOrderReduction/component/contact/MORUnilateralInteractionConstraint.h
@@ -158,7 +158,7 @@ public:
 
 
 #if  !defined(MOR_MORUNILATERALINTERACTIONCONSTRAINT_CPP)
-extern template class SOFA_CONSTRAINT_API MORUnilateralInteractionConstraint<defaulttype::Vec3Types>;
+extern template class SOFA_MODELORDERREDUCTION_API MORUnilateralInteractionConstraint<defaulttype::Vec3Types>;
 #endif
 
 } // namespace constraintset
