+++ src/ModelOrderReduction/component/contact/MORpointModel.h
@@ -21,7 +21,7 @@
 ******************************************************************************/
 #ifndef MOR_POINTCOLLISIONMODEL_H
 #define MOR_POINTCOLLISIONMODEL_H
-#include <ModelOrderReduction/initModelOrderReduction.h>
+#include <ModelOrderReduction/config.h>
 
 #include <SofaMeshCollision/PointModel.h>
 #include <SofaMeshCollision/PointModel.inl>
@@ -79,7 +79,7 @@ protected:
 
 
 #if  !defined(MOR_POINTCOLLISIONMODEL_CPP)
-extern template class SOFA_MESH_COLLISION_API MORPointCollisionModel<defaulttype::Vec3Types>;
+extern template class SOFA_MODELORDERREDUCTION_API MORPointCollisionModel<defaulttype::Vec3Types>;
 
 #endif
